refactor: OAuthController를 interface 대신 abstract class로 정의해서 중복 코드 줄이기
service의 종류나 인가 코드를 요청할 URL만 다를 뿐 메소드를 포함한 나머지 부분은 모두 동일해서 중복 코드를 줄이고자 abstract class를 정의했다.
이 abstract class를 extends하여 service와 요청 URL만 재정의해주니 중복 코드가 현저하게 줄었다.
